slack/pyproject.toml
[project]
name = "klynx-slack-agent"
version = "0.1.0"
description = "KLYNX Slack Agent - receives Slack events, calls Core, posts responses & actions"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.24",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "slack_sdk>=3.27",
  "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools]
packages = ["slack_agent"]

